backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python>=8.3.0
opencv-python>=4.8.0
mediapipe>=0.10.13
numpy>=1.24.3
face-recognition==1.3.0
dlib>=19.24.0
Pillow>=10.1.0
python-dotenv==1.0.0
scikit-learn>=1.3.2
Werkzeug==3.0.1
protobuf>=4.25.3,<5

